better code structure for `node.rs`
