+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["compute", "p2p", "workflows", "utils", "monitor"]
+members = ["compute", "p2p", "workflows", "utils"]
 
 # FIXME: removing this breaks the workflows
 # compute node is the default member, until Oracle comes in
