+++ compute/src/handlers/mod.rs
@@ -1,5 +1,5 @@
 mod pingpong;
 pub use pingpong::*;
 
-mod workflow;
-pub use workflow::*;
+// mod workflow;
+// pub use workflow::*;
