+++ monitor/README.md
@@ -1,5 +1,11 @@
 # Dria Knowledge Network Monitor
 
+> [!WARNING]
+>
+> THIS IS NO LONGER PART OF THE CRATE AS OF VERSION 0.3.
+>
+> IT MAY BE REMOVED SOON.
+
 The monitor node generates a random peer ID, and listens to task messages only. It does not process them or respond to any heartbeat requests. It keeps track of `task` and `result` messages, and prints the "pending" tasks at specific intervals.
 
 ## Usage
